Refactored vertical label out
