+++ .brokk/project.properties
@@ -1,10 +1,10 @@
 #Brokk project configuration
-#Thu Jul 31 15:34:54 CLT 2025
+#Wed Aug 13 13:19:11 SAST 2025
 buildDetailsJson={"buildLintCommand"\:"./gradlew --quiet classes","testAllCommand"\:"./gradlew --quiet test","testSomeCommand"\:"./gradlew --quiet test{{\#classes}} --tests {{value}}{{/classes}}","excludedDirectories"\:[".gradle","project/.bloop",".vscode","bin",".bsp","/nbdist","build-logic/.gradle","project/project","cli/build",".bloop",".direnv","out","target","/dist","/nbbuild","/nbproject/private","/workspace","build","/.nb-gradle",".idea",".metals",".settings"]}
 codeAgentTestScope=WORKSPACE
 code_intelligence_language=JAVA
 code_intelligence_refresh=AUTO
 commitMessageFormat=The commit message should be structured as follows\: <type>\: <description>\nUse these for <type>\: debug, fix, feat, chore, config, docs, style, refactor, perf, test, enh
 dataRetentionPolicy=IMPROVE_BROKK
-issuesProviderJson={"type"\:"GITHUB","config"\:{"kind"\:"github","owner"\:"","repo"\:"","host"\:"","default"\:true}}
+issuesProviderJson={"type"\:"GITHUB","config"\:{"kind"\:"github","owner"\:"","repo"\:"","host"\:""}}
 jiraProjectUserEmail=jbellis@gmail.com
